✨ feat: update import paths for NoticeEventsSection and add NoticeDetailsModal component
