+++ app/page.js
@@ -1,7 +1,7 @@
 import { HeroSection } from "@/components/HeroSection";
 import { AboutSection } from "@/components/AboutSection";
 import { AcademicFeaturesSection } from "@/components/AcademicFeaturesSection";
-import { NoticeEventsSection } from "@/components/NoticeEventsSection";
+import { NoticeEventsSection } from "@/components/NoticeEvents/NoticeEventsSection";
 import { GallerySection } from "@/components/GallerySection";
 
 export default function Home() {
